+++ packages/angular_devkit/build_angular/src/utils/index.ts
@@ -9,6 +9,7 @@
 export * from './default-progress';
 export * from './delete-output-dir';
 export * from './run-module-as-observable-fork';
+export * from './load-proxy-config';
 export * from './normalize-file-replacements';
 export * from './normalize-asset-patterns';
 export * from './normalize-source-maps';
